Enable CORS for app-catalog
This commit enables the apache headers module,
and adds an access control header which will
allow AJAX requests from other sites. This
is intended to allow a Horizon panel to work
with the App Catalog directly.

This commit also includes a modification to
the puppet modeule to create three JSON files from
the contents of YAML files retrieved from the
App Catalog repository.

Change-Id: I4a3c44bb5806362416b1b175d917939ebaa351dd
Closes-Bug: 1467699