Enable CORS for app-catalog
This commit enables the apache headers module,
and adds an access control header which will
allow AJAX requests from other sites. This
is intended to allow a Horizon panel to work
with the App Catalog directly. For multiple
deployments of OpenStack to access the JSON
files regardless of their domain name, this
additional header (Access-Control-Allow-Origin)
must be set to "*" which will enable cross-
origin resource sharing which is normally
forbidden to prevent security issues related
to cross-site scripting.

This commit also includes a modification to
the puppet module to create three JSON files from
the contents of YAML files retrieved from the
App Catalog repository. The YAML format was
chosen in order to make it convenient for humans
to add content to the App Catalog, but the
preliminary work on a Horizon plugin relies on
retrieving the contents in JSON format. Adding
the functionality to automatically generate
the JSON files is a short term step intended
to accelerate development of the Horizon plugin.

Change-Id: I4a3c44bb5806362416b1b175d917939ebaa351dd
Closes-Bug: 1467699