Enable CORS for new /api directory and gzip
This patch adds an access control header to
the location /api/v1/assets.
This will allow a Horizon panel to work with
the App Catalog directly. For multiple
deployments of OpenStack to access the JSON
file regardless of their domain name, this
additional header (Access-Control-Allow-Origin)
must be set to "*" which will enable cross-
origin resource sharing which is normally
forbidden to prevent security issues related
to cross-site scripting. This change also
enables compression on the file, giving roughly
a 5x reduction in download size.

Authored-By: Christopher Aedo <doc@aedo.net>
Co-Authored-By: Kevin Fox <kevin@efox.cc>
Change-Id: I4da417908186c24767310d337bc3bbbc07964d48
Closes-Bug: 1490609