Enable CORS for new /api directory
This patch adds an access control header to
content stored under the /api directory.
This will allow a Horizon panel to work with
the App Catalog directly. For multiple
deployments of OpenStack to access the JSON
files regardless of their domain name, this
additional header (Access-Control-Allow-Origin)
must be set to "*" which will enable cross-
origin resource sharing which is normally
forbidden to prevent security issues related
to cross-site scripting.

Change-Id: I4da417908186c24767310d337bc3bbbc07964d48
Closes-Bug: 1490609