Fix for Murano Liberty version issue.
Murano does not support tagging packages in a way so that they can
add features to a package without making it break with older
engines. The app catalog does not support versioning yet, so we
need a way to let the Liberty Murano find Liberty packages
separate from the Kilo ones. This change sets up a way for the Murano
engine to specify what version it is and we can update the endpoint
later to target Liberty specific Murano packages.

Change-Id: Icb869e9ac8e9795e93d8922600d900522b759e08